requirements.txt
pdfplumber>=0.11.0
watchdog>=4.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
